+++ zholobov.oleg/T1/parallelogram.hpp
@@ -14,11 +14,11 @@ namespace zholobov {
     Parallelogram();
     Parallelogram(const point_t& p1, const point_t& p2, const point_t& p3);
 
-    float getArea() override;
+    double getArea() override;
     rectangle_t getFrameRect() override;
     void move(point_t p) override;
-    void move(float dx, float dy) override;
-    void scale(float k) override;
+    void move(double dx, double dy) override;
+    void scale(double k) override;
   };
 
 }
