+++ zholobov.oleg/T1/shape.hpp
@@ -10,11 +10,11 @@ namespace zholobov {
   class Shape {
    public:
     virtual ~Shape() = default;
-    virtual float getArea() = 0;
+    virtual double getArea() = 0;
     virtual rectangle_t getFrameRect() = 0;
     virtual void move(point_t p) = 0;
-    virtual void move(float dx, float dy) = 0;
-    virtual void scale(float k) = 0;
+    virtual void move(double dx, double dy) = 0;
+    virtual void scale(double k) = 0;
   };
 
 }
