+++ zholobov.oleg/T1/square.cpp
@@ -4,6 +4,6 @@ zholobov::Square::Square():
   Rectangle()
 {}
 
-zholobov::Square::Square(const point_t& a, float length):
+zholobov::Square::Square(const point_t& a, double length):
   Rectangle(a, point_t{a.x + length, a.y + length})
 {}
